+++ src/calendar/day/basic/index.tsx
@@ -1,17 +1,40 @@
 import _ from 'lodash';
 import PropTypes from 'prop-types';
+
 import React, {Component, Fragment} from 'react';
 import {TouchableOpacity, Text, View} from 'react-native';
+
 import {shouldUpdate} from '../../../component-updater';
 import styleConstructor from './style';
 import Marking from '../marking';
 
+interface BasicDayProps {
+  state?: 'selected' | 'disabled' | 'today';
+  /** The marking object */
+  marking: any;
+  /** Date marking style [simple/period/multi-dot/multi-period]. Default = 'simple' */
+  markingType: MarkingTypes;
+  /** Theme object */
+  theme: Object;
+  /** onPress callback */
+  onPress: (date: Object) => void;
+  /** onLongPress callback */
+  onLongPress: (date: Object) => void;
+  /** The date to return from press callbacks */
+  date: Object;
+  /** Disable all touch events for disabled days. can be override with disableTouchEvent in markedDates*/
+  disableAllTouchEventsForDisabledDays: boolean;
+  /** Test ID*/
+  testID: string;
+  /** Accessibility label */
+  accessibilityLabel: string;
+}
 
-export default class BasicDay extends Component {
+export default class BasicDay extends Component<BasicDayProps> {
   static displayName = 'IGNORE';
 
   static propTypes = {
-    state: PropTypes.oneOf(['disabled', 'today', '']), //TODO: deprecate
+    state: PropTypes.oneOf(['selected', 'disabled', 'today', '']),
     /** The marking object */
     marking: PropTypes.any,
     /** Date marking style [simple/period/multi-dot/multi-period]. Default = 'simple' */
@@ -28,23 +51,27 @@ export default class BasicDay extends Component {
     disableAllTouchEventsForDisabledDays: PropTypes.bool
   };
 
-  constructor(props) {
-    super(props);
-    
-    this.style = styleConstructor(props.theme);
-  }
+  style = styleConstructor(this.props.theme);
 
-  shouldComponentUpdate(nextProps) {
-    return shouldUpdate(this.props, nextProps, ['children', 'state', 'markingType', 'marking', 'onPress', 'onLongPress', 'date']);
+  shouldComponentUpdate(nextProps: BasicDayProps) {
+    return shouldUpdate(this.props, nextProps, [
+      'children',
+      'state',
+      'markingType',
+      'marking',
+      'onPress',
+      'onLongPress',
+      'date'
+    ]);
   }
 
   onPress = () => {
     _.invoke(this.props, 'onPress', this.props.date);
-  }
-  
+  };
+
   onLongPress = () => {
     _.invoke(this.props, 'onLongPress', this.props.date);
-  }
+  };
 
   get marking() {
     let marking = this.props.marking || {};
@@ -69,6 +96,10 @@ export default class BasicDay extends Component {
     return disableTouch;
   }
 
+  isSelected() {
+    return this.marking.selected || this.props.state === 'selected';
+  }
+
   isDisabled() {
     return typeof this.marking.disabled !== 'undefined' ? this.marking.disabled : this.props.state === 'disabled';
   }
@@ -90,10 +121,10 @@ export default class BasicDay extends Component {
   }
 
   getContainerStyle() {
-    const {customStyles, selected, selectedColor} = this.props.marking;
+    const {customStyles, selectedColor} = this.marking;
     const style = [this.style.base];
 
-    if (selected) {
+    if (this.isSelected()) {
       style.push(this.style.selected);
       if (selectedColor) {
         style.push({backgroundColor: selectedColor});
@@ -101,7 +132,7 @@ export default class BasicDay extends Component {
     } else if (this.isToday()) {
       style.push(this.style.today);
     }
-    
+
     //Custom marking type
     if (this.isCustom() && customStyles && customStyles.container) {
       if (customStyles.container.borderRadius === undefined) {
@@ -114,10 +145,10 @@ export default class BasicDay extends Component {
   }
 
   getTextStyle() {
-    const {customStyles, selected, selectedTextColor} = this.props.marking;
+    const {customStyles, selectedTextColor} = this.marking;
     const style = [this.style.text];
 
-    if (selected) {
+    if (this.isSelected()) {
       style.push(this.style.selectedText);
       if (selectedTextColor) {
         style.push({color: selectedTextColor});
@@ -138,14 +169,14 @@ export default class BasicDay extends Component {
 
   renderMarking() {
     const {theme, markingType} = this.props;
-    const {selected, marked, dotColor, dots, periods} = this.marking;
+    const {marked, dotColor, dots, periods} = this.marking;
 
     return (
       <Marking
         type={markingType}
         theme={theme}
         marked={this.isMultiDot() ? true : marked}
-        selected={selected}
+        selected={this.isSelected()}
         disabled={this.isDisabled()}
         today={this.isToday()}
         dotColor={dotColor}